third_party/blink/web_tests/platform/generic/fast/dom/shadow/form-in-shadow-expected.txt

PASS obj.hidden is 'hidden'
PASS obj.text is 'text'
PASS obj.checkbox1 is 'on'
PASS obj.checkbox2 is undefined.
PASS obj.range is '50'
PASS obj.textarea is 'textarea'
PASS successfullyParsed is true

TEST COMPLETE
